Add test for no blocking rule scenario
